Added getup.html
When loadBreak loads an activity that requires standing up, a link takes the user to getup.html, which should contain the exact same message, along with a prompt to stand up. The breaktime.html page only contains standing up activities. The link to it appears only if loadBreak() returns the get off your laptop message.