Changed a lot of stuff
Replaced 2 breaktime messages
Added .html to each local href
Moved the noooo function into the event listener

Moved all question redirect messages to the neutral, positive, or negative arrays. Also added a couple more redirect messages.

Fixed bugs for when a randomMessage function is inserted into the console.

Changed the code for the morning animation, but it acts the same.

Added a feature where if a whoa message starts with PLAY, the link changes to a music link.