This is a big one
I've updated some themes. Added some links to a few pages. Adjusted the text of some links.

Page 2 no longer has a 1 in 8 chance to redirect you to Page 52. Instead, it has a 1 in 10 chance to show up a pop-up reminding the user to drink water.

Page 7 (the math game) has been changed a lot. The number pad to choose stuff is between the heading and the bringit message. As a joke, I added a toddler level. The most important update is choosing 0 for number of answers to level up. If selected, the user cannot level up or down. It also changes the way the checkmarks and x's are displayed.

In the script, I replaced all the document.getElementById statements with $, given that I added the function $(x).