Added a new sound effect
Whenever a message with a cross or a radioactive sign shows up, pressing on the "Click Here" button will play a sound and then redirect you to the website. This also applies to the 404 page and Page 5 of CYOA.

This fixes the "Do Nothing For 2 Minutes" bug where audio won't play.