Changed everything
Removed the style tags from the 404 page and replaced it with an href to the stylesheet using https. Modified the stylesheet as a result. Also removed the alerts from the 404 page.

Added some comments to the script page.

Put some arrays outside of the functions. Added a function that adds a whole lot of messages to the redirectMessage array. This includes modified breaktime messages, messages from whoa.html, relapse.html, and repent2.html, some rejected messages that came back,  quotes from the quotes page (cyoa/15.html), and a few new messages. This meant that there are now 224 random messages, including a duplicate message.

Because of this, I added a new page called "haha.html", which contains the original 30 messages.

Also added the function that shrinks text if there's too much of it and the function where the ROFL clip plays when there's a ROFL emoji.