Updated pages
Added a new genre to the cyoa/58.html page.

Made it so that the webpages with names of websites can load two functions. The getBack() function has a 1 in 30 chance of replacing the page contents with the infamous "Get back to whatever you should be doing!" and "Immediately close this tab" text.

In some of the changed pages, pressing "No" will persuade the user "Don't do nothing. Do something!" If the user presses "No" again, they'll be taken to different pages.

If they're on the YouTube page, it will redirect them to a page asking them what they wanted to watch.

If they're on the Newgrounds page, it will redirect them to a page asking them about spare time.

The reddit and Facebook pages act normal, but the DeviantArt page requires pressing "No" three times instead of two to achieve the same results.

Added a new event listener where pressing "M" in the webpage will take them to the music slider page, but only iff the music pop-up appears.

Also, added a few more links regarding music break/redirect messages.