Added and changed stuff.
On the YouTube page, the "Don't do nothing. Do something!" link goes to a new page called breakyt.html, which tells you to do things that feel more rewarding than just browsing YouTube videos and finding recommended videos that aren't really that rewarding.

On the Newgrounds page, the webpage predicts that you wanted to play a game after saying "No" once. It acts similarly to the deviantart page.

Rearranged the loadWebsite numbers

Added a new goal to get a drawing tablet to draw art digitally.

Rearranged the Something CYOA page (cyoa/11.html) by adding the music page.

CYOA page 20 now gives different links depending on if the page is accessed through mobile or not, because I refuse to pay $9.99/month for Spotify.

CYOA page 40 is now different. If "Don't feel like checking" is pressed on Page 2, it takes you to Page 40, which tells the user how to access the gTasks list. As a result, clicking "My Day is Done!" redirects to Page 41 instead.