Added a bunch of stuff
Expanded the default array to 34 messages.

Added a new break message and disabled a few break messages related to homework.

Added a new link that is put in the break pages and redirects you to the brave page. It is desined to be the end all be all of motivational pages.