Deprioritized Eurovision, Added a new table to cyoa.28
I changed a few files so that watching Eurovision won't be considered as a break.

Also on CYOA Page 28, I added another element. This time, it's a table listing things ordered by priority. As a result, the table stylings had to be adjusted, and the javascript code on page 28 had to be adjusted.