Made major changes
Enhanced the website pages, added a new script called interactivefiction.js for those pages, and replaced the dropbox links with insighttimer links for guided meditation.

Also added an old YouTube video to the server for when the user wants to laugh or something.

The new features include typing in the phrase "do something" to take a break, the word "focus" taking the user to page 18 in cyoa, any word inciding fears, anger, aggression, and calmness bringing up a random link to one of Andrew Johnson's meditations, any word inciting humor showing the Sileny Ota YouTube video, and some other keywords that provides a link to page 2 in CYOA with each message providing a different message depending on the word.