Added the busy page and a new function
Sometimes, I say that I'm busy, and one of the website pages will take me to the busy page. This is designed for when I'm busy doing something. It will prompt the user to not overwork themself, take them to the music page, not weeb out, or anything from the interactivefiction script, most notably prompting the user for an energy boost.

The new function is designed to replace the for loop with the if condition returning true for saving space.

Also on DeviantArt, a responce for typing in Touhou is added.