Created a page, updated a branch
I created the anime utilities page where I can have a convenient place to open up all my most frequently used anime utilities, including the new anime timer.

On the busy page, I replaced the nextanime link with the animeutilities link.

On the interactivefiction script, a phrase that contains anime will have the same effect as a phase that contains lazy on the pages that uses the interactive fiction script except for deviantart.