This is a big one, especially for the google page
For the website pages, the dropdown menus are formed using Javascript arrays and loops.

The Google page gets a big makeover.  Depending on the response, a slider may show up, "Yes", "No", and "Maybe" may show up, or the appropriate text appears like normal. Instead of asking for emotions, the page asks for the importance of the search term. A variety of responses could be made based on the importance level selected. This includes an urgency drop down menu.

There's also a new page for a new branch: The disappointed Kirby page.

Also, arrays can be put in as a parameter.

Modified the cyoa stylesheet for the YouTube search page and the Rock Paper Scissors game.

Modified the root stylesheet to add a slider and to create a new class for the relax page,