Major changes to the break page
If you don't want to do something, it will ask you why. The page it takes you to depends on the reason. Sometimes, the page will just reload.

The brave and the new 2 minute rule page also contains the "But I can't" excuse. It takes you to cyoa page 42.

Also on the interactive fiction script is a new response to hungry if it's fast Sunday.