Create goals.html
This is a big one. This page is meant to keep track of my goals.

There is a grid showing the habits in a randomized order. These habits also have a duration of time. When a grid element is clicked on, it may take you to a relevant page. Each grid element has a randomly generated color and a corresponding letter. If a keyboard letter is pressed, then that habit "disappears", kind of like checking off an item off a to-do list.

If all the items are checked off, the webpage congratulates the user for doing them all, as the page cycles through some colors. It won't be easy, but it can be done in a single day.