Replaced coninue with continue, added a new function called cute(number)
The new function arises when the user is weebing out. It asks three questions. Answer yes to all of them, and the webpage brings the link to the nextanime page, which will contain anime I've rewatched such as GochiUsa, Ro-Kyu-Bu!, and Gonna Be the Twin-Tail!! Maybe another imouto anime would be added to the mix.