Improved the timer
moveTime no longer adds event listeners.

The setTime function on the start button is now the startTimer function. It passes parameters to the setTimer function.

The console.log for times are now more meaningful.

Instead of using onclick in JavaScript, removeEventListener and addEventListener are used to switch the pauseTimer and resumeTimer functions.