Expanded the relax2 page
Now there's a whole lot of meditations on that page. The "I'm Feeling Lucky" button on that page will only take the user to any of the top 16 most popular meditations.

Depending on the time of day, the Power Nap or the Evening Boost meditation may be unavailable through the "I'm Feeling Lucky" button.