The deviantart page now has its own "Let me guess" feature
When you go to the DeviantArt page, the webpage guesses that you're thinking about cute anime girls. If you say "Yes", or you just started thinking about them, a slider is presented. The slider is used to gauge how much the thought of cute anime girls controls the user. The response depends on the slider value. The highest value will provide a link to the Tabernacle Choir music stream.

Fixed the styling of the sliders. Some are thin, while others are thick.

Changed the link on cyoa/8 to breaktime instead of bigbreak

Added a new five minute timer webpage that starts instantly. It is designed to get rid of unnecessary thoughts. After five minutes, it asks the user what their thoughts are like.