Modified Goals, Added New Pages for New Goals
I modified goals so that they can be aligned spiritually, because I've been neglecting God. It doesn't matter if being a member of the Church of Jesus Christ of Latter-Day Saints will cause social rejection, because I think being an active member in that regard will lead to high success.