Added stylesheet, added a new font
The reason why so many files are being affected is because I added a new stylesheet to each page.

From what I learned from skillshare, I figured that I should be using the Barlow Font instead of the Futura and Avenir fonts. That way, the website should look consistent for most people who see this website.