Added a new array for the goals list
I added things that I wanted to do. The A habits (A-F) are what I feel are essential. The B habits (G-L) are what I feel to be good. The C habits (M-R) are what I feel to be fun distractions.

Also, changed the month for reading journal entries to June 2012.