Added a new feature to the goals list
The export feature exports the goal list to plain text in a textarea. This can be used for Notepad or Google Sheets.

The reason for making an export feature is because I want to track my goals, even on mobile, and this was the best I could do.

Also, I made the text of the goals list smaller for screens larger than 1135px wide.