Pandora Navigator is here!
Now, each a in the grid has the "hfire-link" class. Each grid item with a higher intensity is marked with a red plus. Each grid item with a lower intensity is marked with a blue minus.

If Page A has a link to Page B, then Page B must link to Page A. If Page A has a link to Page C with a plus, then Page C must link back ot Page A with a minus.