Changed the styles
I decided to change the colors of nine themes since theme1 reminded me of a certain website too much. The old styles are moved to a new stylesheet that's unused.

Other documents that depend on colors of the nine themes have been affected to, such as cyoa/47 and the goals page easter egg.