Added a lot of stuff
There is now a games folder for the two games I created on this website and the short games link.

I also improved the experience for the games selection. Now there's reference to exergaming, and it's designed towards playing upstairs.

Added a new quote for cyoa/15.html

cyoa/20.html now has YouTube links, and cyoa/64.html, which was redirected from cyoa/20.html depending on the size of the screen, was deemed unnecessary

cyoa/55.html now links to cyoa/1.html instead of cyoa/4.html

The youtube page now has the label "Other Reason" to avoid confusion.