Overhauled how the website pages worked
Each website now displays a big message like the "haha" page and has an id of "message".  As a result, the text formerly in "message" moves to "question", and those pages now use the script.js script along with the interactivefiction.js script. Most id's in the interactivefiction script have been changed as a result, and some functions in multiple scripts had to be renamed, modified, or deleted.

Also in the interactivefiction script, some branches have been added, moved, removed, and modified. The patient branch has been merged. Some branches check to make sure that the middle text doesn't match the top text.

The script.js script has been changed a lot to accomodate for the new webpages, especially at the randomLink function.

I fixed a feature where if nothing is inputted in the google or youtuberesults page, it will go straight to the importance slider. I also rearranged the slider values a little.

I also fixed the sileny ota easter egg on the YouTube page.

Also, I discovered max-width, an essential part to responsive design. It makes it so that pages don't appear funky on some pages on mobile.