Adjusted the anime slider and goals, fixed a couple titles
Adjusted the result for "I really want to look at cute anime girls". Now it displays a prompt to be more creative instead. If it's rejected, then it displays a prompt to be more religious instead. If that's rejected, the page will be disappointed in you and will link you to the anti-procrastination page.

The relax/meditate goal now requires 2 sessions to be successful.

The titles of gcprogress and cyoa/17 have been changed.