Added the objective.html page, updated breaktime.js
For when you feel like your life or day has no objectives, you can go to the objective.html page and have the website give you an objective for the day for you to do. The objectives are grouped into tiers. With the creation of the new page, there needed to be ways to link to the objective page.

The "No me gusta abburido!" link in cyoa/13.html now takes the user to page 54, which asks if the the user has any goals or objectives for today. One link leads to the relax page with the lazy query. The second link leads to the complete tasks page (AKA 2.html). The third leads to cyoa/64.html. In that page, the user can choose betwen 1 objective, 10 goals, 1 random thing, or 8 random things.

The breaktime.js script now contains a link to the objective script if the message "Make an objective for today." shows up. I also decided to change "I don't feel like it" to two links: "I already did" and "It's impossible". "I already did" reloads the page. "It's impossible" gives three links. "I don't want to do it" takes the user to either the brave or 2minuterule page. "Actually, I'll do it" undoes the action. "Circumstances out of my control prevent me from doing it" reloads the page. A lot of code had to be commented out.

Other changes
1.html: "Random Stuff" now links to time.html.
3.html & 4.html: "Now now. Maybe later." is changed to "I can't do it now." Link is unchanged.
6.html: "I can't think of anything to do." now links to 64.html.
11.html: "Unknown" now links to 64.html.
17.html: The "Enjoy your lazy, unproductive day" function now has a new line containing a link to objective.html, telling the user to choose a C-Tier Objective for having a lazy, unproductive day.
cyoa/style.css: Switched the fallback body fonts.
interactivefiction.js: deleted a branch, merged two branches into one, and added new text to the new branch. It contains two links to 13.html and 54.html.
relax.html: There's now a button called "I'm lying down" to replace "More". The links are the same.
script.js and searchterm.html: Replaced "multiple" value with "really bored" value.
script.js: The "really bored" branch replaces the multiple branch and contains two links to 13.html and 54.html. It also removes the "small" class from the link element if the page is the youtube page.
style.css: Added the link.small tag and changed the body font from Futura Md BT to Futura MdCN BT.
youtube.html: If the word "procrastinator" describes the user, they are taken to the anti-procrastination page. There is a new option called "bored", which executes the "really bored" branch of the finalTest function in interactivefiction. Also made some styling modifications by making some text smaller.