Update emergency.html
Instead of the phrase "Take a cold shower, or drive your car to DragonForce music", a search query shows up with three buttons to search for stuff. These buttons don't work as intended on purpose, and the user has to explain why they are searching for it.

The "Take a cold shower, or drive your car to DragonForce music", message still has a chance to appear at the end of the uncontrollable urge path.