Revamped the old emergency button, updated the morning emergency page
The old emergency button has now changed. After clicking on "Emergency", a search bar shows up, and the user can search for anything they want. Except, that none of the buttons work as predicted, and the user is asked why they wanted to search for it. After the user submits the reason, the webpage will deny the search, tell you one of 30 reasons why searching for it is bad, and bring back the Emergency button, which works almost like it used to. Depending on the message, the user might be taken to a religious page, or they might be given a command to play "Fury of the Storm".

morning/emergency: The colored squares should work as intended, and one button is more religous than the others. I also moved the cold showers text to a new webpage. Also made a few text changes.