Changed a link, made the site a little more mobile friendly
animeslider/deviantart: Changed the link that appears in response to "I want to look at cute anime girls" in animeslider.html and deviantart.html. It now links to bigbreak.html.

breaktime.js: Rearranged the switch case branches, and added a new break message. Also fixed a bug and made it so that it's easier for mobile users to go to the mobile page.

style.css: Added new rules to reduce the number of columns of the three-list and four-list class on smaller screens.

time.html: Now redirects immediately to breakm.html if the screen size is small enough.