Updated relax2 and cyoa/2, fixed styling
The title is no longer randomized upon going to the page, but there is a "Random Meditation" button. New meditations were added along with a new category. New moods were also added.

Some links leading to relax2 have been updated with moods.

In 2.html, a link to 64.html is added.

Style fixes: Smallest max-width is 413px.
Added a new rule for buttons with the relaxing theme.