Quadrupled number of morning goals pages
The original morining goals page is now changed, as it takes inspiration from cyoa/17.html. Now factors include if it's raining outside or if today would be a good day to do the laundry.

Originally I was going to make the goals page interactive by making elements swappable or by making a cyoa page, but that would require a lot of effort.