interactiveFiction.js now asks how you're feeling instead of why you aren't drawing
The reason why I did this is because the statement "Why don't you start drawing?" creates a lot of resistance. So, I changed the question to hopefully reduce the amount of resistance to draw something.

Also updated the sick branch, the music branch, the tired branch with words implying bloatedness, the want and need branches, the bored and really bored branches, and other branches inloving expressing desire for cuteness

Also on the deviantart page, saying "Yes" will just automatically submit the word "cute", which puts the ineffective anime slider out of use.