Update emergency.html
Sometimes, I resist the temptation to search for something in that search bar, because it might be a better option. So, I added an option to resist the temptation to search, and the webpage praises me for doing that. I also added in a new button called "Do something else"