Added the sleep subdirectory
The sleep directory is designed to be used at 10:00PM before going to sleep. It is designed to wrap up the day. Features include asking the user if their day is done in the index page, instructions on how to go to sleep in the done page, some things to do in the not-done page, as well as a searchbar and a page similar google.html.

Also fixed a typo. "OK! Add that to your "Search Terms" list to search for later!"