Added more stuff
Exercise is essential, and I will not walk outside in very cold weather. So, I decided to substitute walking for Stepmania. Added a new special morning Stepmania page as a result.

Changed "Face Mask" to "Sleep Mask" and "Tired" to "Tired and Lazy" in the morning pages.

I added question marks to almost every 15.html link so that it can be more personalized. I plan on improving it later.

In 20.html, I added a new button called skip. It just shows an alert persuading the user to listen to the clip.

In 55.html, the two links lead to 54.html and 65.html, with "Been there, done that" leading to page 65, asking the user what time it is. There are five options. Three of them lead to a Stepmania page, one of them leads to the relax2 page, and one leads to the sleep page.

I added a new tired page, inspired by Better Ideas's video "Why you're always tired". It is linked by the struggle2 page, but there are plans to link it with more pages.