Made the website more supportive of art goal
I've basically avoided my main goal of drawing, so I decided to change some stuff in the Interactive Fiction and Breaktime scripts. This led to the art, artstage, artstruggle pages to be updated and the creativity page to be, well, created. A new General Conference talk has been watched as a result.