Improved the atomic-habits page, added a new deviantart page
The deviantart page has proven to be ineffective, so now there's a new deviantart-habits page that uses the same script from the atomic-habits page. Except, most of the quotes inspire the killing of bad habits like going to DeviantArt on autopilot.

The deviant-art page also contains functions from the interactive-fiction script.