Created a buzzfeed-style quiz, updated styling
I created my own buzzfeed-style quiz and updated the styling on grids. A way to access this quiz is through the interactivefiction or breaktime scripts, or through the goals page.

I also updated some stylin, in regards to the grid, and now I'm finding pages that should be changed to make the pages more consistent.