Created the atomic habits table
This webpage pulls from the atomic habits quotes array and makes a table. A page number array is added to the script to clarify page numbers. It also features alternating colors to make the tables easier to read.

Added new style rules for the new webpage
Added new button to the atomic habits page