Added more features to the timer
Biggest new feature is the addition of the "Things to do during break" grid. I also added URL search parameters so that the timer can start in the 2 hour position when the 30 minute or 1 hour timer is stopped. And when the 2 hour timer is stopped, it goes to the nextanime page.

Any time anything in the anime grid in nextanime is clicked on, two tabs open. One is the anime page, and the other is the anime timer.