Made tons of improvements, changed things around
For the "Listen to this." And "Clear your mind." Suggestions, switched to a Notion database of Andrew Johnson meditations with the quick filter enabled. The text will be changed to "Develop the Andrew Johnson meditations database."

Added a new break and a new C-Tier objective: Watch Only Connect

Added a new break and a new C-Tier objective: Make blog posts

Adjusted the links for General Conference talks to a Dynalist link.

Commented out anything that has to do with college or finding a job.