Update nextanime.html
Forgot to take out the lines that makes rolling a 1 and 9 guaranteed.

Anyway, how did I prioritize the Banished anime with code? When you roll the dice for the first time, you are asked if you watched Banished from the Hero's Party. If you answer yes, the first roll will always be a 3. Now that's rigged dice!