Added localStorage to breaktime pages
Before the changes, the current breaktime message will disappear if the webpage becomes inactive. Now, the current breaktime message will stay even if the webpage becomes inactive or it gets closed.

The displayBreak function is split into two parts, with displayBreak mostly dealing with the DOM and setBreak dealing with things like loading the current breaktime message or generating a random breaktime message.